isolate the NodeLabels type into a dedicated source
