+++ pkg/eks/api/api.go
@@ -8,8 +8,6 @@ import (
 	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
 	"github.com/aws/aws-sdk-go/service/eks/eksiface"
 	"github.com/aws/aws-sdk-go/service/sts/stsiface"
-	"strings"
-	"errors"
 )
 
 const (
@@ -193,36 +191,3 @@ type (
 	// NodeLabels labels nodes via kubelet's --node-labels flag
 	NodeLabels map[string]string
 )
-
-func (f NodeLabels) String() string {
-	s := ""
-	for k, v := range f {
-		if s != "" {
-			s += ","
-		}
-		s += k + "=" + v
-	}
-	return s
-}
-
-func (f *NodeLabels) Set(value string) error {
-	if *f == nil {
-		*f = map[string]string{}
-	}
-	kvs := strings.Split(value, ",")
-	for i := range kvs {
-		kv := strings.Split(kvs[i], "=")
-		if len(kv) != 2 {
-			return fmt.Errorf("node label must be formatted K=V, but it was: %s", kvs[i])
-		}
-		if kv[0] == "" {
-			return errors.New("key must not be omitted in --node-labels")
-		}
-		(*f)[kv[0]] = kv[1]
-	}
-	return nil
-}
-
-func (f *NodeLabels) Type() string {
-	return "NodeLabels"
-}
